move event/site to own files/cleanup
